requirements.txt
anyio==3.6.2
fastapi==0.95.1
greenlet==2.0.2
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing-extensions==4.5.0
